+++ src/UglyToad.PdfPig/DocumentLayoutAnalysis/MathExtensions.cs
@@ -22,7 +22,7 @@ namespace UglyToad.PdfPig.DocumentLayoutAnalysis
         /// Computes the mode of a sequence of decimal values.
         /// </summary>
         /// <param name="array">The array of decimal.</param>
-        public static decimal Mode(this IEnumerable<decimal> array)
+        public static double Mode(this IEnumerable<double> array)
         {
             return array.GroupBy(v => v).OrderByDescending(g => g.Count()).First().Key;
         }
