+++ src/UglyToad.PdfPig/Fonts/CidFonts/Type2CidFont.cs
@@ -14,8 +14,8 @@
     {
         private readonly ICidFontProgram fontProgram;
         private readonly VerticalWritingMetrics verticalWritingMetrics;
-        private readonly IReadOnlyDictionary<int, decimal> widths;
-        private readonly decimal? defaultWidth;
+        private readonly IReadOnlyDictionary<int, double> widths;
+        private readonly double? defaultWidth;
         private readonly CharacterIdentifierToGlyphIndexMap cidToGid;
 
         public NameToken Type { get; }
@@ -35,8 +35,8 @@
         public Type2CidFont(NameToken type, NameToken subType, NameToken baseFont, CharacterIdentifierSystemInfo systemInfo,
             FontDescriptor descriptor, ICidFontProgram fontProgram,
             VerticalWritingMetrics verticalWritingMetrics,
-            IReadOnlyDictionary<int, decimal> widths, 
-            decimal? defaultWidth,
+            IReadOnlyDictionary<int, double> widths,
+            double? defaultWidth,
             CharacterIdentifierToGlyphIndexMap cidToGid)
         {
             Type = type;
@@ -51,11 +51,11 @@
             this.cidToGid = cidToGid;
 
             // TODO: This should maybe take units per em into account?
-            var scale = 1 / (decimal)(fontProgram?.GetFontMatrixMultiplier() ?? 1000);
+            var scale = 1 / (double)(fontProgram?.GetFontMatrixMultiplier() ?? 1000);
             FontMatrix = TransformationMatrix.FromValues(scale, 0, 0, scale, 0, 0);
         }
 
-        public decimal GetWidthFromFont(int characterIdentifier)
+        public double GetWidthFromFont(int characterIdentifier)
         {
             if (fontProgram == null)
             {
@@ -70,7 +70,7 @@
             return GetWidthFromDictionary(characterIdentifier);
         }
 
-        public decimal GetWidthFromDictionary(int characterIdentifier)
+        public double GetWidthFromDictionary(int characterIdentifier)
         {
             if (widths.TryGetValue(characterIdentifier, out var width))
             {
@@ -82,7 +82,7 @@
                 return defaultWidth.Value;
             }
 
-            return Descriptor?.MissingWidth ?? 1000m;
+            return (double)(Descriptor?.MissingWidth ?? 1000);
         }
 
         public PdfRectangle GetBoundingBox(int characterIdentifier)
